+++ src/commands/hardis/lint/unusedmetadatas.ts
@@ -47,8 +47,10 @@ export default class UnusedMetadatas extends SfdxCommand {
   };
   /* jscpd:ignore-end */
   protected outputFile: string;
+  protected outputFilesRes: any = {};
   // Comment this out if your command does not require an org username
   protected static requiresUsername = false;
+  protected static supportsUsername = true;
   // Comment this out if your command does not support a hub org username
   protected static supportsDevhubUsername = false;
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
@@ -78,9 +80,10 @@ export default class UnusedMetadatas extends SfdxCommand {
     }
 
     if (unusedLabels.length > 0 || unusedCustomPermissions.length > 0) {
+      await this.buildCsvFile(unusedLabels, unusedCustomPermissions);
       const branchMd = await getBranchMarkdown();
       const notifButtons = await getNotificationButtons();
-
+      globalThis.jsForceConn = this?.org?.getConnection(); // Required for some notifications providers like Email
       NotifProvider.postNotifications({
         type: "UNUSED_METADATAS",
         text: `Unused metadatas detected in ${branchMd}\n`,
@@ -89,7 +92,6 @@ export default class UnusedMetadatas extends SfdxCommand {
         severity: "warning",
         sideImage: "flow",
       });
-      this.buildCsvFile(unusedLabels, unusedCustomPermissions);
     } else {
       uxLog(this, "No unused labels detected or custom permissions detected.");
     }
@@ -189,6 +191,6 @@ export default class UnusedMetadatas extends SfdxCommand {
       ...unusedCustomPermissions.map((permission) => ({ type: "Custom Permission", name: permission })),
     ];
 
-    await generateCsvFile(csvData, this.outputFile);
+    this.outputFilesRes = await generateCsvFile(csvData, this.outputFile);
   }
 }