+++ src/commands/hardis/org/diagnose/unusedlicenses.ts
@@ -70,6 +70,7 @@ export default class DiagnoseUnusedLicenses extends SfdxCommand {
 
   protected debugMode = false;
   protected outputFile;
+  protected outputFilesRes: any = {};
   protected permissionSetLicenseAssignmentsActive = [];
   protected permissionSetLicenses = [];
   protected unusedPermissionSetLicenseAssignments = [];
@@ -176,7 +177,7 @@ export default class DiagnoseUnusedLicenses extends SfdxCommand {
     // Generate output CSV file
     if (this.unusedPermissionSetLicenseAssignments.length > 0) {
       this.outputFile = await generateReportPath("unused-ps-license-assignments", this.outputFile);
-      await generateCsvFile(this.unusedPermissionSetLicenseAssignments, this.outputFile);
+      this.outputFilesRes = await generateCsvFile(this.unusedPermissionSetLicenseAssignments, this.outputFile);
     }
 
     // Manage notifications
@@ -326,12 +327,14 @@ export default class DiagnoseUnusedLicenses extends SfdxCommand {
 
       const orgMarkdown = await getOrgMarkdown(this.org?.getConnection()?.instanceUrl);
       const notifButtons = await getNotificationButtons();
+      globalThis.jsForceConn = this?.org?.getConnection(); // Required for some notifications providers like Email
       NotifProvider.postNotifications({
         type: "UNUSED_LICENSES",
         text: `${unusedPermissionSetLicenseAssignments.length} unused Permission Set Licenses Assignments have been found in ${orgMarkdown}`,
         attachments: [{ text: notifDetailText }],
         buttons: notifButtons,
         severity: "warning",
+        attachedFiles: this.outputFilesRes.xlsxFile ? [this.outputFilesRes.xlsxFile] : [],
       });
     }
     return [];