+++ src/commands/hardis/org/test/apex.ts
@@ -140,6 +140,7 @@ You can override env var SFDX_TEST_WAIT_MINUTES to wait more than 60 minutes
           // Send notification
           const orgMarkdown = await getOrgMarkdown(this.org?.getConnection()?.instanceUrl);
           const notifButtons = await getNotificationButtons();
+          globalThis.jsForceConn = this?.org?.getConnection(); // Required for some notifications providers like Email
           NotifProvider.postNotifications({
             type: "APEX_TESTS",
             text: `Apex Tests run coverage issue in ${orgMarkdown}\nTest run coverage ${coverageTestRun}% should be > to ${minCoverageTestRun}%`,
@@ -172,12 +173,14 @@ You can override env var SFDX_TEST_WAIT_MINUTES to wait more than 60 minutes
       // Send notification
       const orgMarkdown = await getOrgMarkdown(this.org?.getConnection()?.instanceUrl);
       const notifButtons = await getNotificationButtons();
+      globalThis.jsForceConn = this?.org?.getConnection(); // Required for some notifications providers like Email
       NotifProvider.postNotifications({
         type: "APEX_TESTS",
         text: `Apex Tests are failing in ${orgMarkdown}`,
         attachments: [{ text: testResultStr }],
         buttons: notifButtons,
         severity: "error",
+        attachedFiles: fs.existsSync(reportDir + "/test-result.txt") ? [reportDir + "/test-result.txt"] : [],
       });
       // (LEGACY) Send notification if possible
       if (await canSendNotifications()) {