+++ src/common/utils/index.ts
@@ -1169,3 +1169,7 @@ export async function isMonitoringJob() {
   }
   return false;
 }
+
+export function getNested(nestedObj, pathArr) {
+  return pathArr.reduce((obj, key) => (obj && obj[key] !== "undefined" ? obj[key] : undefined), nestedObj);
+}