+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [4.28.0] 2024-02-21
+
+- Minimum Node.js version is now v18
+- New notifications channel: **EmailProvider** (use variable **NOTIF_EMAIL_ADDRESS** that can contain a comma-separated list of e-mail addresses)
+- Update existing call to notifications to add attached files when Email notif channel is active
+- Audit trail suspect actions: add the number of occurences for each suspect action found
+- Add more not suspect actions: dkimRotationPreparationSuccessful,createdReportJob,deletedReportJob,DeleteSandbox
+- Get tickets info: also check in ticket ids in branch name
+- Remove force config restDeploy=true
+- Rename _Provided by sfdx-hardis_ into _Powered by sfdx-hardis_
+
 ## [4.27.1] 2024-02-10
 
 - Skip post-deployment notifications if nothing was deployed in delta mode